Move Player.py from game_map.py to player.py
